http://tools.yalamanchili.info:8080/browse/SSTO-1724 update leave request error fix
git-svn-id: http://tools.sstech.us/svn/office/trunk@4397 04aae24c-3eb2-d842-8c26-b4285eb26770

+++ office-master/office-bpm/src/main/java/info/yalamanchili/office/bpm/OfficeBPMService.java
@@ -53,11 +53,13 @@ public class OfficeBPMService {
     }
 
     public void deleteProcessInstance(String processInstanceId) {
-        try {
+        if (findProcessInstance(processInstanceId) != null) {
             bpmRuntimeService.deleteProcessInstance(processInstanceId, "user cancled");
-        } catch (Exception e) {
-            e.printStackTrace();
         }
+    }
+
+    public ProcessInstance findProcessInstance(String processInstanceId) {
+        return bpmRuntimeService.createProcessInstanceQuery().processInstanceId(processInstanceId).singleResult();
     }
 
     public String getDeployedProcessInfo() {
